mv_clip/mv_clip_lvlm/requirements.txt
numpy
tqdm
pillow
transformers
wandb
scikit-learn